requirements.txt
spotipy==2.18.0

psycopg2-binary

python-dotenv==0.19.0

pandas==2.1.4

sqlalchemy-redshift


apache-airflow==2.2.3
